refactor: tests structure
