--- __tests__/unit/example.test.ts
@@ -1,5 +1,0 @@
-describe("some test", () => {
-  it('is just a test', () => {
-    expect(1).toEqual(1)
-  })
-})
